Refactor and update multiple components of the project
- Modifications in the profile and word controllers.
- Updates to email and profile services.
- Changes in interfaces and models related to words and responsible entities.
- Removal of the abstract_db.v file.
- Renaming of the user model adapter to word model adapter.
- Adjustments in repositories and entities for users and words.

These changes aim to enhance the structure and functionality of the system.

+++ src/infra/user/entities/user_temp.v
@@ -8,7 +8,7 @@ pub struct UserTemp {
 pub:
 	first_name        string
 	last_name         ?string
-	responsible       Responsible @[sql: int; sql_type: 'smallint']
+	responsible       Responsible @[sql: int; sql_type: 'int']
 	birth_date        time.Time
 	email             string
 	password          string
